{cmake} Fix deprecation warning by updating min version to 3.5
With CMake 3.27, we get a deprecation warning:

"Compatibility with CMake < 3.5 will be removed from a future version of CMake.
Update the VERSION argument <min> value or use a ...<max> suffix to tell
CMake that the project does not need compatibility with older versions."

Bump min version to 3.5 to remove warning.